Fixed most compiler warnings in cross-builds to djgpp and mingw.
git-svn-id: file:///home/survex-svn/survex/trunk@996 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/caverot.c
@@ -853,7 +853,7 @@ static void
 parse_command(int argc, char **argv)
 {
    int c;
-   int col_idx = 0;
+   size_t col_idx = 0;
 
    cmdline_set_syntax_message("3D_FILE...", NULL); /* TRANSLATE */
    cmdline_init(argc, argv, short_opts, long_opts, NULL, help, 1, -1);
